Deleted duplicated tests
basic_tests.py contained all the unit tests in meterage_tests.py, plus
more.  Therefore I removed meterage_tests.py, moved basic_tests.py to
meterage_tests.py, and now am committing.
